chore: clean up project structure and organize assets
