+++ public/images/hero/logos/bitcoin-search.png
